requirements.txt
boto3==1.28.38
opencv-python==4.8.0.76
pillow==10.0.0
python-dotenv==1.0.0